automate/requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dj-database-url==0.5.0
Django==2.2.8
django-crispy-forms==1.9.1
django-decouple==2.1
django-filter==2.3.0
django-heroku==0.3.1
django-rq==2.3.2
django-sniplates==0.7.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
install==1.3.3
lxml==4.5.2
Markdown==3.2.2
paramiko==2.7.1
postgres==3.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycurl==7.43.0.5
PyNaCl==1.4.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
rq==1.4.3
six==1.15.0
sqlparse==0.3.1
suds-jurko==0.6
urllib3==1.25.9
whitenoise==5.2.0
zipp==3.1.0
celery==4.4.7
django-celery-results==1.2.1
django-celery-beat==2.0.0